requirements.txt
PyYAML==3.12
pytest==3.0.2
